op2: init: Reduce swappiness to 15
Oh dear... 100 is really ludicrous since we only need to compress
memory via zRAM in dire cases. compressing memory for the heck of
it is taxing on battery life and other resources such as CPU cycles.

I can see what was trying to be done here, but constantly compressing
memory won't help with reasources at all.